# public/res/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "An api for interacting with an E-Commerce server.\n\n### Admin user info:\n##### Username | 'admin'\n##### Password | 'password'",
    "version": "1.0.0",
    "title": "Online E-Commerce API",
    "contact": {
      "email": "arigorn15@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://localhost:3000/",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/cart": {
      "summary": "The user's cart information",
      "description": "Operate on the logged in user's cart.",
      "get": {
        "summary": "Get user's cart information",
        "description": "Request the information for the logged in user's cart.",
        "operationId": "get_cart",
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "products": [
                    {
                      "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                      "name": "example_product1",
                      "price": "$5.00",
                      "qty": 1,
                      "image_file": "image-1709469725513.jpg"
                    }
                  ],
                  "total_price": 5.24
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          }
        },
        "tags": ["Cart"]
      },
      "post": {
        "summary": "Add items to cart",
        "description": "Add products to the logged in user's cart, and return the changed cart information. Returns updated cart.",
        "operationId": "post_cart",
        "requestBody": {
          "description": "Product Information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cart_update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "products": [
                    {
                      "id": 1,
                      "name": "example_product1",
                      "price": 5.0,
                      "qty": 1,
                      "image_file": "image-1709469725513.jpg"
                    },
                    {
                      "id": 2,
                      "name": "example_product2",
                      "price": 5.0,
                      "qty": 1,
                      "image_file": "image-1709469725513.jpg"
                    }
                  ],
                  "total": 10.0
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          }
        },
        "tags": ["Cart"]
      },
      "put": {
        "summary": "Update item quantity in cart",
        "description": "Updates the quantity of a product in the logged-in user's cart, and return the changed cart information.",
        "operationId": "put_cart",
        "requestBody": {
          "description": "Product Information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cart_update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "products": [
                    {
                      "id": 1,
                      "name": "example_product1",
                      "price": 5.0,
                      "qty": 1,
                      "image_file": "image-1709469725513.jpg"
                    },
                    {
                      "id": 2,
                      "name": "example_product2",
                      "price": 5.0,
                      "qty": 2,
                      "image_file": "image-1709469725513.jpg"
                    }
                  ],
                  "total_price": 15.0
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          },
          "404": {
            "description": "Cart item not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 500,
                  "message": "Can't edit product.  Does not exist in cart!"
                }
              }
            }
          }
        },
        "tags": ["Cart"]
      },
      "delete": {
        "summary": "Delete item from cart",
        "description": "Delete products from the logged in user's cart, and return the changed cart information.",
        "operationId": "delete_cart",
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "required": true,
            "schema": {
              "type": "uuid"
            },
            "description": "The id of the product to delete from the user's cart"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "products": [
                    {
                      "id": 1,
                      "name": "example_product2",
                      "price": 5.0,
                      "qty": 2,
                      "image_file": "image-1709469725513.jpg"
                    }
                  ],
                  "total_price": "$5.00"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          },
          "404": {
            "description": "Cart item not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 500,
                  "message": "Can't edit product.  Does not exist in cart!"
                }
              }
            }
          }
        },
        "tags": ["Cart"]
      }
    },
    "/cart/checkout": {
      "summary": "Retrieve payment intent",
      "description": "Add a payment intent to the user's cart.",
      "post": {
        "summary": "Create a payment intent",
        "description": "Add a stripe payment intent (retrieved from the /secret endpoint) to the logged in user's cart in the database.",
        "operationId": "post_checkout",
        "requestBody": {
          "description": "Payment Intent",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/checkout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "message": "payment intent pi_3MtwBwLkdIwHu7ix28a3tqPa created!"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "text/plain; charset=utf-8": {
                "example": "Please log in!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "status": 500,
                  "message": "Something went wrong!"
                }
              }
            }
          }
        },
        "tags": ["Cart", "Checkout"]
      }
    },
...
    "/orders": {
      "summary": "Order information",
      "description": "View all orders",
      "get": {
        "summary": "Get list of orders",
        "description": "Search the database for and return all orders made by the current user",
        "operationId": "search_orders",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": false,
            "schema": {
              "type": "uuid"
            },
            "description": "The id of the user to retrieve orders for.  If user is not an admin, only their own orders will be returned."
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "date": "2023/02/24",
                    "total": "$15.00",
                    "shipping_status": "delivered",
                    "payment_status": "success"
                  },
                  {
                    "id": 2,
                    "date": "2022/04/27",
                    "total": "$100.00",
                    "shipping_status": "shipped",
                    "payment_status": "success"
                  },
                  {
                    "id": 3,
                    "date": "2022/03/05",
                    "total": "$35.00",
                    "shipping_status": "pending",
                    "payment_status": "success"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "No orders found for user with id <userId>!"
                }
              }
            }
          }
        },
        "tags": ["Orders"]
      }
    },
    "/orders/{orderId}": {
      "summary": "Order information",
      "description": "View specific order",
      "get": {
        "summary": "Get an order's details by id",
        "description": "Search the database for and return the order with the given id.",
        "operationId": "search_order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "The id of the user's order.",
            "required": true,
            "schema": {
              "type": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                  "date": "2023/02/24",
                  "total": 15.0,
                  "shipping_status": "delivered",
                  "payment_status": "success",
                  "tracking_number": "123456789",
                  "products": [
                    {
                      "id": 1,
                      "name": "example_product",
                      "price": 5.0,
                      "qty": 2,
                      "image_file": "image-1709469725513.jpg"
                    },
                    {
                      "id": 2,
                      "name": "example_product2",
                      "price": 5.0,
                      "qty": 1,
                      "image_file": "image-1709469725513.jpg"
                    }
                  ],
                  "shipping_address": {
                    "addr_line_1": "123 Main St",
                    "addr_line_2": "Suite 1",
                    "city": "San Diego",
                    "state": "CA",
                    "zip_code": 92101
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "text/plain; charset=utf-8": {
                "example": "Order with ID <orderId> not found."
              }
            }
          }
        },
        "tags": ["Orders"]
      }
    },
    "/orders/admin": {
      "summary": "Order information",
      "description": "View all orders",
      "get": {
        "summary": "Get list of orders",
        "description": "Search the database for and return all or filtered orders for all users",
        "operationId": "search_admin_orders",
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "type": ["string", "uuid"]
            },
            "description": "Either the shipping status, or id of the order or user."
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "date": "2023/02/24",
                    "total": "$15.00",
                    "shipping_status": "delivered",
                    "payment_status": "success"
                  },
                  {
                    "id": 2,
                    "date": "2022/04/27",
                    "total": "$100.00",
                    "shipping_status": "shipped",
                    "payment_status": "success"
                  },
                  {
                    "id": 3,
                    "date": "2022/03/05",
                    "total": "$35.00",
                    "shipping_status": "pending",
                    "payment_status": "success"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please log in!"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "No orders found!"
                }
              }
            }
          }
        },
        "tags": ["Orders"]
      }
    }
  },
  "components": {
    "schemas": {
...
      "cart_update": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "uuid"
          },
          "qty": {
            "type": "integer"
          }
        },
        "xml": {
          "name": "Cart_Update"
        }
      },
      "checkout": {
        "type": "object",
        "properties": {
          "payment_intent": {
            "type": "string"
          }
        }
      },
...
    }
  }
}